+++ varfish/static/js/qc_plots.js
@@ -162,7 +162,7 @@ function colorFgToBg(col) {
 function plotlyRelatedness(response, containerId) {
   const relData = response.data.relData
   if (relData.length === 0) {
-    $("#" + containerId + " .placeholder-container .placeholder-text").html('<span class="text-muted"><i class="fa fa-ban fa-3x"></i><br /><p class="pt-3 font-italic">No relatedness plot available for a singleton.</p></span>');
+    $("#" + containerId).remove()
     return;
   }
   const clsRelData = {